refactor(add-transaction): split the sell/withdraw workflow into SellForm (#467)
Follow-up to the buy-forms split (#495). Extract the sell/withdraw section — the
shared holding-picker + amount shell with per-asset inserts (fund units, bank
received, gold qty/price) and per-asset summaries — into `SellForm` in
addTransactionForms.tsx (~260 lines of JSX).

The sheet now computes the SellPreview once as an object and passes it to SellForm
(which destructures every derived figure it renders); it keeps only the ten values
it still needs for the submit gate, the buildSellPayload call, and the gold-price
prefill. Sell input state/setters + holding selection are passed as props;
business math stays in addTransactionModel.ts. Now-unused imports (LoadError,
number-format helpers, three lucide icons) drop out of the sheet. No behavior or
markup change.

Parity is guarded by the existing AddTransactionSheet component tests (which drive
the sell flow for fund/bank/gold and assert the exact withdrawal POST bodies) plus
the add-transaction + book-withdrawal E2E — all green.

Verified: AddTransactionSheet 24 + SellWithdrawSheet 16 + full unit suite 1250
passed; typecheck 0 errors; lint 0 errors; E2E add-transaction (14) +
book-withdrawal (4) passed. AddTransactionSheet drops 959 → 703 lines (1189 → 703
across both split PRs).
